Clean-up the slides ☕
