+++ src/slides/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import Introduction from './introduction'
+import ProceduralGeneration from './procedural-generation'
+import StarSystem from './star-system'
+
+export default () => (
+  <div className='slides'>
+    <Introduction />
+    <ProceduralGeneration />
+    <StarSystem />
+  </div>
+)
